rename: Move TextWithNewlines to common components
